+++ src/components/TaskList/TaskList.jsx
@@ -3,7 +3,7 @@ const { connect } = require('react-redux')
 
 const TaskListItem = require('../TaskListItem')
 const hookUpStickyNav = require('../hookUpStickyNav')
-const Filters = require('../../models/filters')
+const Filters = require('../../models/Filters')
 const LastFilter = require('../../models/last-filter')
 
 class TaskList extends React.Component {
